requirements.txt
streamlit
ipywidgets
matplotlib
numpy
pandas
torch
torchvision
tqdm
Pillow
face-detection==0.1.1
opencv-python-headless==4.9.0.80